requirements.txt
appdirs==1.4.3
asgiref==3.2.10
beautifulsoup4==4.9.1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1
django-rest-registration==0.5.6
djangorestframework==3.11.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
typing==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
